fix: update profile
